filtered empty line from input
IT WAS THE DAMN NEW LINE.

For ONCE I forgot to filter for the new line and this happened.

I had the right solution for a long time but this damn new line threw my code totally out of whack and caused what looked like an off by one at first. GRRRRRRRRRRRRRRRRRR